% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundary}
\alias{boundary}
\title{Boundary}
\format{
A sf polygon object with 1 row and 1 variable
}
\usage{
data(boundary)
}
\description{
An sf polygon object, that limits the area where polygons will be extended.
}
\details{
\itemize{
  \item geom. sfc_POLYGON information
}
}
\keyword{datasets}
