% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conglomerate.R
\name{conglomerate}
\alias{conglomerate}
\title{Function that creates the new conglomerate}
\usage{
conglomerate(A, weigth, sl, idp = NULL)
}
\arguments{
\item{A}{An incidence matrix.}

\item{sl}{integer with lower limit of the size of the primary sampling units.}

\item{weight}{Number of occupied housing units in the blocks}

\item{id}{string with ID of new Primary Unit Sampling - PSU}
}
\description{
ESPECIFICAR TODO LOS DETALLES TËCNICOS NECESARIOS
PUEDEN USARSE VARIAS LINEAS E INTRODUCIR ECUACIONES
\deqn{\hat{\sigma}^2_y = \frac{\sum_s w_k (y_k - \bar{y}_w)^2}{\sum_s w_k}}
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas.
Valliant, R, et. al. (2013), \emph{Practical tools for Design and Weighting Survey Samples}. Springer
}
\author{
Javier Nuñez <Javier_Nunez at inec.gob.ec>

Christian Garces <Christian_Garces at inec.gob.ec>
}
