% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_weights}
\alias{df_weights}
\title{Dataset with number of household in every block}
\format{
A data frame with 249 PSU rows and 2 variables:
\describe{
  \item{id}{id of PSI}
  \item{viv}{Number of households in every household}
}
}
\source{
{Created as example of wieghts}
}
\usage{
df_weights
}
\description{
Dataset with number of household in every block
}
\keyword{datasets}
