% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ext_pol}
\alias{ext_pol}
\title{Extended polygons}
\format{
A sf multi-polygon object with 249 rows and 2 variables.
}
\usage{
data(ext_pol)
}
\description{
A sf multi-polygon object, geographic data of adjacent polygons generally
related to blocks. ext_pol is created by running the polygon_ext function
with the parameters set in the example.
}
\details{
\itemize{
  \item id. a character string, blocks ID related to poly argument, preferably unique by block.
  \item geom. sfc_MULTIPOLYGON information.
}
}
\keyword{datasets}
