% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extpol.R
\name{extpol}
\alias{extpol}
\title{Function that extends the polygons based on cartographic information}
\usage{
extpol(poli, profile, id = NULL, boundary = NULL, buf = 5, density = 0.1)
}
\arguments{
\item{poli}{Non-adjacent polygons generally refer to blocks.}

\item{profile}{The polygon that defines the area where the extension work will take place}

\item{id}{ID of the original Primary Sampling Units}

\item{boundary}{It's a multi-polygon that represents rivers or boundaries, used to avoid connections between blocks separated by them, and it can or cannot not be contained within the profile}

\item{buf}{Buffer Width is a scalar value that determine the quality of the polygon extension. Higher quality requires more processing}

\item{density}{TDensity buffer is a scalar value that determine the quality of the polygon extension. Higher quality requires more processing}
}
\description{
This function allows for extending the polygons prior to the formation of Primary Sampling Units
}
\details{
ESPECIFICAR TODO LOS DETALLES TËCNICOS NECESARIOS
PUEDEN USARSE VARIAS LINEAS E INTRODUCIR ECUACIONES
\deqn{\hat{\sigma}^2_y = \frac{\sum_s w_k (y_k - \bar{y}_w)^2}{\sum_s w_k}}
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas.
Valliant, R, et. al. (2013), \emph{Practical tools for Design and Weighting Survey Samples}. Springer
}
\author{
Javier Nuñez <Javier_Nunez at inec.gob.ec>

Christian Garces <Christian_Garces at inec.gob.ec>
}
