% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gap}
\alias{gap}
\title{Gap}
\format{
A sf multi-polygon object with 1 row and 1 variable.
}
\usage{
data(gap)
}
\description{
A multi-polygon sf object, geographic data that represents rivers or
boundaries, used to avoid undesired connections between blocks, could be
contained within the boundary.
}
\details{
\itemize{
  \item geom. sfc_POLYGON information
}
}
\keyword{datasets}
