% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{im}
\alias{im}
\title{Incidence matrix}
\format{
A matrix object with 249 rows and 249 columns.
}
\usage{
data(im)
}
\description{
A matrix object that represents the binary relation 'polygon i shares a
common boundary with polygon j'. im is created by running the inc_matrix
function with the parameters set in the example.
}
\details{
\itemize{
  \item the ith row represents the ith polygon of ext_pol.
  \item the jth column represents the jth polygon of ext_pol.
}
}
\keyword{datasets}
