% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_matrix.R
\name{inc_matrix}
\alias{inc_matrix}
\title{Function that builds the incidence matrix (IM)!}
\usage{
inc_matrix(poly, tol = 10, id = NULL)
}
\arguments{
\item{poly}{a sf polygon object, extended polygons that share a common boundary.}

\item{tol}{an integer, tolerance in meters.}

\item{id}{a character string, the column name of the ID related to extended polygons, preferably unique by polygon.}
}
\value{
IM
}
\description{
This function returns a binary matrix that represents the binary relation 'polygon i shares a common boundary
with polygon j'.
}
\details{
Incidence matrix is computed by the geographical relations between a set of adyacent polygons. The binary
relation is derived by intersecting all the polygons with each other. As a result, only line types polygons
are expected.

If the intersection length between polygon i and polygon j is greater or equal than tol, then the IM[i, j] = IM[j,i] = 1.

Its important to mention that diag(IM) = 0.
}
\examples{
inc_matrix(poly, tol=10, id=id)
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas.
Valliant, R, et. al. (2013), \emph{Practical tools for Design and Weighting Survey Samples}. Springer
}
\author{
Angel Gaibor <mat.angel.gaibor at gmail.com>

Javier Núñez <mat.javier.nunez at gmail.com>
}
