% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incmatrix.R
\name{incmatrix}
\alias{incmatrix}
\title{Function that build incidence matrix}
\usage{
incmatrix(pol, tol = 10, id = NULL)
}
\arguments{
\item{pol}{Non-adjacent polygons generally refer to blocks.}

\item{tol}{Tolerance in meters.}

\item{id}{ID of the original Primary Sampling Units}
}
\description{
The binary relation that represents the incidence matrix is 'Polygon A shares a border with Polygon B.
It returns a matrix with 0 and 1 that specify if there is incidency between poligons
}
\details{
ESPECIFICAR TODO LOS DETALLES TËCNICOS NECESARIOS
PUEDEN USARSE VARIAS LINEAS E INTRODUCIR ECUACIONES
\deqn{\hat{\sigma}^2_y = \frac{\sum_s w_k (y_k - \bar{y}_w)^2}{\sum_s w_k}}
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas.
Valliant, R, et. al. (2013), \emph{Practical tools for Design and Weighting Survey Samples}. Springer
}
\author{
Javier Nuñez <Javier_Nunez at inec.gob.ec>

Christian Garces <Christian_Garces at inec.gob.ec>
}
