% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly}
\alias{poly}
\title{Non-adyacent polygons}
\format{
A sf multi-polygon object with 249 rows and 2 variables
}
\usage{
data(poly)
}
\description{
An sf multi-polygon object, geographic data of non-adjacent polygons generally
related to blocks.
}
\details{
\itemize{
  \item id. a character string, blocks ID related to poly argument, preferably unique by block.
  \item geom. sfc_MULTIPOLYGON information.
}
}
\keyword{datasets}
