% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_ext.R
\name{polygon_ext}
\alias{polygon_ext}
\title{Function that extends the polygons based on geographic data related to blocks.}
\usage{
polygon_ext(poly, boundary, id = NULL, gap = NULL, buff = 5, density = 0.1)
}
\arguments{
\item{poly}{an sf polygon object, geographic data of non-adjacent polygons generally related to blocks.}

\item{boundary}{an sf polygon object, that limits the area where polygons will be extended.}

\item{id}{a character string, the column name of the ID related to blocks in poly argument, preferably unique by block.}

\item{gap}{an multi-polygon sf object, geographic data that represents rivers or boundaries, used to avoid undesired connections between blocks, could be contained within the boundary.}

\item{buff}{an integer (buffer width) that determine the quality of the polygon extension. Higher quality requires more processing power.}

\item{density}{an integer (number of points by 1 meter) that determine the quality of the polygon extension. Higher quality requires more processing power.}
}
\description{
This function allows for extending the polygons prior to the formation of Primary Sampling Units.
}
\details{
polygon_ext is a function that extends non-adjacent polygons until they share a common boundary.

The objective of the extension is to identify which polygons share a common boundary, facilitating
the delimitation of Primary Sampling Units, which are generally composed of adjacent polygons sets.

The extension is carried out by generating Voronoi Polygons from a inner points mesh constructed
from the buff and density parameters. Higher values require greater computational processing capacity,
however they can drastically improve the quality of the result.

Using the gap parameter, unwanted incidences between extended polygons can be eliminated (for example,
large avenues or rivers that run through the city).
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas.
Valliant, R, et. al. (2013), \emph{Practical tools for Design and Weighting Survey Samples}. Springer
}
\author{
Angel Gaibor <mat.angel.gaibor at gmail.com>

Javier Núñez <mat.javier.nunez at gmail.com>
}
