% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psu_clustering.R
\name{psu_clustering}
\alias{psu_clustering}
\title{Function that creates Primary Sampling Units from adjacent polygons.}
\usage{
psu_clustering(A, W, lowerLimit, idp = NULL, weight)
}
\arguments{
\item{A}{an n_n integer matrix, incidence matrix.}

\item{W}{a two column data.frame, contains the id and weights of the adyacent polygons (generally the id and the number of occupied dwellings in each block).}

\item{lowerLimit}{an integer, defines the lower limit of the PSU size to be created.}

\item{idp}{an character, name of id column in W.}

\item{weight}{a character, name of weights column in W.}
}
\description{
This function returns a data.frame object that identifies the adyacent polygons
along with the PSU to which they belong.
}
\details{
The psu_clustering functions takes as inputs an A_{n_xn} incidence matrix and a w_n
weight vector (subset of W) related to the n adyacent polygons in A. While creating Primary
Sampling Units (PSU), generally w_n takes the number of occupied dwellings in the n
block (adyacent polygon) as weight. A_{n_xn} and w_n must be ordered in the
same way.

The function returns a vector of identifiers of size n that relates the polygons to
the PSU which it belongs.
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas.
Valliant, R, et. al. (2013), \emph{Practical tools for Design and Weighting Survey Samples}. Springer
}
\author{
Angel Gaibor <mat.angel.gaibor at gmail.com>

Javier Núñez <mat.javier.nunez at gmail.com>
}
