% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuR-package.R
\docType{package}
\name{psuR-package}
\alias{psuR}
\alias{psuR-package}
\title{psuR: Build Primary Sampling Units fom Geographic Data Related to Blocks.}
\description{
A structured process to build Primary Sampling Units (PSU) from geographic data linked to blocks specially orientated to household surveys.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matjaviernunez/psuR}
  \item Report bugs at \url{https://github.com/matjaviernunez/psuR/issues}
}

}
\author{
\strong{Maintainer}: Javier Núñez \email{mat.javier.nunez@gmail.com}

Authors:
\itemize{
  \item Angel Gaibor \email{mat.angel.gaibor@gmail.com}
}

Other contributors:
\itemize{
  \item Andrés Gutiérrez \email{andres.gutierrez@cepal.org} [contributor]
  \item José Zea \email{jfzeac@unal.edu.co} [contributor]
}

}
\keyword{internal}
