% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights}
\alias{weights}
\title{Block weights}
\format{
A data.frame object with 249 rows and 2 variables
}
\usage{
data(weights)
}
\description{
A two column data.frame, contains the id and weights of the adyacent polygons
 (generally the id and the number of occupied dwellings in each block).
}
\details{
\itemize{
  \item idp. a character string, blocks ID, preferably unique by block.
  \item weight. a integer, weights of adyacent polygons.
}
}
\keyword{datasets}
